[SERV-3] Update product and package services

+++ src/main/java/org/vladimirskoe/project/service/ProductServiceImpl.java
@@ -1,13 +1,12 @@
 package org.vladimirskoe.project.service;
 
-import org.vladimirskoe.project.entity.Product;
-import org.vladimirskoe.project.dao.ProductRepository;
-
-import org.springframework.stereotype.Service;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.vladimirskoe.project.exception.NullObjectException;
-
 import java.util.List;
+import java.util.Optional;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.vladimirskoe.project.dao.ProductRepository;
+import org.vladimirskoe.project.entity.Product;
+import org.vladimirskoe.project.exception.NullObjectException;
 
 @Service
 public class ProductServiceImpl implements ProductService {
@@ -19,31 +18,38 @@ public class ProductServiceImpl implements ProductService {
         this.productRepository = productRepository;
     }
 
+    @Override
     public Product addProduct(Product product) {
-        if (product == null) {
-            throw new NullObjectException("Input value is null and can't be added. ");
-        } else {
-            return productRepository.addProduct(product);
-        }
+        return productRepository.addProduct(product);
     }
 
+    @Override
     public Product getProductById(Integer id) {
-        return productRepository.getProductById(id);
+        return productRepository
+                .getProductById(id)
+                .orElseThrow(() -> new NullObjectException("Product not found"));
     }
 
+    @Override
     public List<Product> getAllProducts() {
         return productRepository.getAllProducts();
     }
 
+    @Override
     public Product updateProduct(Integer id, Product product) {
-        if (product == null) {
-            throw new NullObjectException("Input value is null and can't be updated.");
-        } else {
-            return productRepository.updateProduct(id,product);
+        Optional<Product> optionalProduct = productRepository.getProductById(id);
+        if (!optionalProduct.isPresent()) {
+            throw new NullObjectException("Product does not exist and cannot be updated");
         }
+        return productRepository.updateProduct(product);
     }
 
+    @Override
     public void deleteProduct(Integer id) {
+        Optional<Product> optionalProduct = productRepository.getProductById(id);
+        if (!optionalProduct.isPresent()) {
+            throw new NullObjectException("Product does not exist and cannot be deleted");
+        }
         productRepository.deleteProduct(id);
     }
 }
